Disable tests failing when built against libc++
